Package project as deployable CLI
