+++ .env.example
@@ -1,0 +1,4 @@
+API_URL=http://192.168.50.4:8787/infer
+NOTES_DIR=/Users/you/Library/Mobile Documents/com~apple~CloudDocs/LLM/Index LLM
+PER_REQUEST_DELAY_SECONDS=120
+DELAY_BETWEEN_CALLS_SECONDS=30
