+++ .gitignore
@@ -1,0 +1,7 @@
+__pycache__/
+*.pyc
+.env
+.DS_Store
+dist/
+build/
+*.egg-info/
