+++ .pre-commit-config.yaml
@@ -1,0 +1,7 @@
+repos:
+- repo: https://github.com/astral-sh/ruff-pre-commit
+  rev: v0.5.6
+  hooks:
+    - id: ruff
+      args: [--fix]
+    - id: ruff-format
