+++ src/obsidian_note_gen/__init__.py
@@ -1,0 +1,5 @@
+"""Obsidian note generator package."""
+
+__all__ = []
+
+__version__ = "0.1.0"
